Added xslTransformerFactory in testEA_Docx.xml
Added missing targetParameter xslTransformerFactory in
testEA_Docx.xml. Without this parameter, an error is given by
ShapeChange:
"Parameter 'outputFormat' contains 'DOCX' and/or 'FRAMEHTML'.
These formats require an XSLT 2.0 processor, which should be
set via the configuration parameter 'xslTransformerFactory'.
That parameter was not found, and the default TransformerFactory
implementation is not 'net.sf.saxon.TransformerFactoryImpl'
(which is known to be an XSLT 2.0 processor); ensure that the
parameter is configured correctly.