improves facet detection, adds documentation to derivedDocumentation()
This commit addresses #53 by searching for simple type mappings in the
direct or indirect supertypes of a basic type, to identify the XSD type
that should be used in the check if a particular facet is supported.
The commit also addresses #54, by adding 'documentation' as a possible
replacement in InfoImpl.derivedDocumentation(...).