add ProfileLoader transformation
A new transformation has been added as part of the profiling
enhancements: the ProfileLoader. It supports loading profiles from a set
of external models (format: SCXML) into the input model of the
transformation. The transformation supports validation of profiles
(input model, loaded model, resulting model). It also supports computing
and logging differences between a loaded model and the input model.