update to ProfileTransferEA and ModelExport
ProfileTransferEA has been improved to only load required information
from the target EA repository, not the whole model. Target EA repository
can now be specifically configured in target configuration, so is no
longer restricted to be the input model of the whole ShapeChange
workflow.
ModelExport now supports a parameter to match the name of tagged values
to ignore in the export (e.g. 'profiles').