improve descriptors
When processing very large models, the previous descriptor
implementation had memory issues (overhead due to duplicate strings and
small/sparse collections). String interning is now applied for
LangStrings and the values of a descriptor are stored in a
DescriptorValue object, with improved memory consumption (only if
multiple LangStrings apply to a descriptor will a list be used to store
them; the dominant case is that a descriptor only has a single
LangString as value).