improve logging of unknown msg number, update package fullNameInSchema
MessageSources usually use a message like "unknown message" as default
value. The class name of the message source has been added to these
default messages to make it easier to find logging issues (where an
unknown message number is used). 
The way that fullNameInSchema for a package is computed has been updated
to avoid returning null. If the package has no target namespace, then
the fullName is returned. Otherwise fullNameInSchema is computed as
usual.