fix property postprocessing and profile validation
Current postprocessing and validation routines of PropertyInfoImpl do
not apply to properties of basic types, code lists, and enumerations.
This check has now been moved from ModelImpl.postprocessClass() (which
was the wrong place for that check) to
PropertyInfoImpl.postprocessAfterLoadingAndValidate().
When checking that the profiles of a value type contain the profiles of
a property, profile validation did not take into account that the
property may inherit the profiles from its class. This has been fixed.