fixes for SQL notEncoded and default value quoting
When marking a class that participates in an association as notEncoded,
the association roles are automatically not encoded as well. That has
implications for the relationship (that is represented by the
association) as a whole: if the class is not mapped, then the whole
relationship is dropped (now a warning is logged in such a case),
otherwise we proceed as usual.

When an attribute has an initial value, and the value type is a code
list or enumeration, the default value is automatically quoted (this can
be overridden via the according map entry param characteristic).

The commit also provides updates to documentation in
ReplicationSchemaConstants and SqlConstants.