remove 'final' on XxxImpl methods, open TagAlias, remove Model.type()
A number of XxxImpl methods could be final, and were set accordingly.
However, that caused issues in a project that depends on ShapeChange.
Thus, we eventually agreed to remove the 'final' again to support
extending ShapeChange, and add a WARNING in the javadoc of these
methods.

The TagAlias configuration element only supported mapping of tags to
well-known descriptor tags. The restriction in the ShapeChange
configuration has been removed, to support mapping of tagged values.

Also a slight change to how the target of a TypeMapEntry is loaded by
the Ontology target: before it split on all occurrences of a colon, now
it only splits on the first.

Method Model.type() has been removed, also the model constants in
Options.java. Occurrences where those constants were checked have been
replaced with instanceof checks.