fix/workaround mixin issue
Mixins were not correctly handled when the EA model was copied into a
GenericModel. The reason was that before the GenericModel was created,
the Converter had already invoked Options.reset(). This also removed the
rules defined by all targets, which should be ok, but caused problems.

For the execution of the first transformers, we need to have the rules
from all targets in Options.java. The reason is that some checks are
performed that try to match on specific target conversion rules. An
example is method ClassInfoEA.baseClass(), which at some point may need
to match on"rule-xsd-cls-mixin-classes-non-mixin-supertypes". When
creating a copy from an EA model, it would be problematic if the rules
configured by the targets were no longer known. Thus, we now reset the
fields of Options.java only after the GenericModel copy has been
created. Eventually, what we really want is to avoid dependencies on
target conversion rules when loading a model. Target specific model
checks should be performed separately. Refactoring them into target
specific model validation classes may be the way to go.