Add error message when a default encoding rule is specified but not configured
Add an error message in the result when a default encoding rule is specified
for a certain platform but this default encoding rule is not present in the ShapeChange
configuration. This adds useful information for the user, as this misconfiguraton
likely will result in other errors or issues later on.