Update and add rules in XSD target; update ShapeChangeResult
XSD target:

- Renamed rule-xsd-cls-featureType-gmlsf-byReference to
rule-xsd-prop-featureType-gmlsf-byReference, also updated the
implementation of the rule: it is now executed as one of the special
cases when adding a property.
- Updated how base and base type are computed in
rule-xsd-prop-constrainingFacets.
- Added rule-xsd-prop-metadata-gmlsf-byReference: Applies simple
byReference encoding of properties with metadata as value type
(determined by tagged value _isMetadata_ on the property, and map
entries in the ShapeChange configuration).

ShapeChangeResult:

- Extended message that warns about invalid constraints during
postprocessing phase of a transformation.
- Added flag and method to determine if a fatal error has been reported.