Update PODS SQL rule, minor changes to XSD target and OCL parsing
XML Schema target:
- Absence of 'xsdDocument' info (from tagged value or PackageInfo
element) logged as warning instead of fatal error. Log message extended
to provide more explanations.

SQL target:
- Update to pods specific code list conversion rule: codeStatusCL and
codeStatusNotes columns instead of activeIndicatorLF and sourceCL
columns.

OCL handling:
- Update to check of
rule-xsd-cls-codelist-constraints-codeAbsenceInModelAllowed: code list
or enumeration with missing enum/code matches this rule or
rule-xsd-all-notEncoded - to cover case in which the class is part of an
external schema that is not encoded.