update flattener (remove name component), update ShapeChangeResult
Flattener update: when applying the transformation to remove name and
code components, the transformation checks if such a change results in a
class with properties having the same name. That fact will be logged as
an error, with details available on INFO level.

Move messages from ShapeChangeResult to specific classes (e.g.
Flattener, Profiler, TransformationManager).