Update CDB target + Flattener transformation, new DescriptorTransformer
CDB target:
- use model element name for CDB attribute symbol / CDB feature name

Flattener transformation:
- Updated and extended merging of descriptors. Now applies to flattening
of multiplicity, complex types, and homogeneous geometries. The
separator to use when merging descriptors (with a suffix or other
descriptor) can now be configured per descriptor. 

new DescriptorTransformer:
- Rule to update descriptors of selected model elements. Selection of
model elements and specification of relevant descriptors to update is
achieved via specific configuration elements, which are contained in the
advanced process configuration.