Do not add import when targetElement not present
In the XML Do not add an import statement, nor a declaration of the XML
namespace abbrevation, when rule rule-xsd-prop-targetElement is not
enabled.

The import statement is not needed when rule-xsd-prop-targetElement is
not enabled. When it is present, and the XML schema that is referred to,
does not exist, it makes the XML validation fail.

A reason for the non-existence of the XML schema can be that the
application schema is maintained by another organation, that uses
another encoding than XML.