Fix unit tests + config validator, update Flattener and SQL target
Unit tests:
- In case of a multi test in dirs, BasicTest now checks that reference
files are actually matched by result files. If a reference file is not
matched by a result file, the test fails.

Configuration validators:
- Fixed a number of validators, to use parameters from the given
configuration if target specific parameters are checked. This also makes
the code simpler to read. Currently, it is not safe to rely on options
for retrieving target parameters during the validation of target
configurations.

General model:
- Added loading of linked documents (for ClassInfo).

SQL DDL target:
- Documentation of Column(s) and Table(s) is now stored in actual
attributes of these classes. Before, documentation was derived
on-the-fly, for example from represented classes and properties. This
change supports explicitly setting the documentation, especially in case
of columns/tables that have no corresponding model element.
- Extended rule-sql-cls-code-lists and rule-sql-cls-code-lists-pods so
that documentation for the code, code status, code notes, and descriptor
columns of a code list table can be configured.