fixed and extended ArcGIS workspace target
- The type of foreign key fields is now determined using the primary key
field of the referenced class (before it was always
esriFieldTypeInteger).
- Replaced HashSets and HashMaps with sorted equivalents where possible,
to avoid issues with different processing order (especially during unit
tests). An example where this plays a role is unique naming of model
elements, since this can depend upon the order in which the model
elements are processed.
- If the class from the application schema has a linked document, then
the linked document will be set for the according class in the ArcGIS
workspace model.