UmlModel target: add stereotype mapping + parameter eapTemplate
Addresses the issue that the stereotypes within a model written by the
UmlModel target can be assigned to the wrong UML Profile / MDG. See #122

Stereotype mappings support mapping of a well-known stereotype to a
stereotype of the intended UML Profile / MDG (by providing a suitable
qualified stereotype, for example YourProfile::featureType).

The parameter eapTemplate can be used to write the model into an EA
repository to which your profile has been loaded (removing the need to
have the profile installed on the computer / the EA environment where
ShapeChange is executed).