fix Schematron creation in case of ISO 19139 encoded value type
XmlSchema target / Schematron generation:
- Fix wrong id attribute when referring to an ISO 19139 encoded type
(requires @id instead of @gml:id).
- Fix treatment of special cases from ISO 19139 that have simple content
and are not referenceable; example is gco:CharacterString. These cases
were incorrectly treated as objects (being referenceable, and not
comparable to a string or number). Introduced new XML attributes in
XsdMapEntry to explicitly state that the XML representation of the type
is not referenceable and that it has simple content.