Update parameter handling of CodelistDictionariesML target
- Parameter 'languages' is no longer required. It defaults to the value
of parameter 'defaultLang'. If a non-multilingual dictionary shall be
created, the two parameters can simply be omitted.
- Configuration validation now checks that a locale reference is given
for each language other than the default language.
- In rule-cldml-prop-codeListAndCodeNameAsGmlId, all characters that
match the regular expression [^a-zA-Z0-9_] are removed from the property
name (to avoid issues if a code value contains special characters that
are not allowed by xs:ID).