Fix ISO19139 union encoding, extend EA utils, fix UmlModel unit test
In the ISO 19139 encoding, a union type is not in the substitution group
of gco:AbstractObject, and is therefore not referenceable.

Added a couple of methods to the EA utility classes.

The UnitTest UmlModelTest.testLinkedDocuments requires the ShapeChange
MDG. If the MDG is only installed locally, and not imported into the
model, then, if the profile was deactivated locally, the test failed.
The MDG has now been imported to the model template.