Update profile constraints trf, XmlSchema tgt, schema selection behavior
Profile constraint transformer:
- Renamed ProfileSchemaConstraintTransformer to
ProfileConstraintTransformer (same for the configuration validator).
- General revisions after additional testing:
  - renamed and added parameters
  - added rule to prohibit base schema types that have a direct
unsuppressed subtype in the profile schema (also influences the rule to
create general out of scope constraints)
  - The transformer will add the tagged value
'prohibitedInProfileSchema' to every property and class that is
prohibited (per OCL constraint added by the transformer). This tagged
value can be used in subsequent processing steps, for example deriving
code list constraints in the XmlSchema target, to prevent creation of
checks (OCL constraints or Schematron assertions) on prohibited model
elements. 
Closes #138
  
XmlSchema target:
- Added rule-xsd-all-propertyAssertion-ignoreProhibited to prevent
creation of a property assertion if either the property or the class for
which the assertion is created is prohibited (currently indicated by
tagged value 'prohibitedInProfileSchema' = true).
- Added parameter 'skipXmlSchemaOutput', with which one can skip writing
XML Schemas. That can be useful if only the Schematron schema is of
interest. Closes #137

Other:
- Schemas selected for processing are now determined by first looking at
the appSchemaNameXxx parameters in the current process configuration (if
set) and otherwise using the parameters from the input configuration.
The generic model no longer stores the IDs of these schemas, and
transformations can no longer explicitly set them. Closes #136
- Added some more map entries for ISO 19139 encoded types.