Remove isSchema from SCXML, handling nav props for SCXML ex- and import
- Remove Package/isSchema element from SCXML. Also removed isSchema as
settable field in class GenericPackageInfo. Whether or not a package is
a schema should be computed dynamically (as implemented by class
PackageInfoImpl), taking into account PackageInfo elements from the
input configuration. Otherwise, PackageInfo elements would be completely
ignored when loading a model from SCXML. UnitTest reference files had to
be updated to remove the isSchema element.
- Added checks to ModelExport and GenericModel reader to ensure that
only navigable properties of a class are exported and read (by a
GenericClassContentHandler). Non-navigable properties (i.e., association
roles) are stored in an association.