Add input parameter applyDescriptorSourcesWhenLoadingScxml
With the new parameter, ShapeChange can be instructed to always load
descriptors from configured sources (e.g. tagged values), when loading
models from SCXML. That supports use cases in which a subset of SCXML is
used (which does not include the descriptors element, but tagged
values).

Descriptor sources are now read from implementations of the Model
interface. ModelImpl provides a default implementation. If a specific
model type has specific sources or defaults for descriptors, then it can
now give this information (this will increase decoupling).

The model export target now only prints the descriptors element if a
model object really has descriptor values.

Also updated documentation of Info.id() and several methods of class
TaggedValues which return the 'first' of a set of values for a tag.

Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>