add boolean isOwned to PropertyInfo interface
This field indicates if an association role is owned by the class at the
other end of the association. Note that this is independent of
navigability. Updated EA and SCXML model classes to read this
information. Updated ModelExport and UmlModel targets to write this
information (for the latter, parameter includeAssociationOwnership is
available to control if the information is actually written).

Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>