Change behaviour length qualifier
Parameter lengthQualifier is always the default, if set, also for fields
with textual data type of limited length that do represent a UML
property. If parameter lengthQualifier is set, it is overridden by
characteristic lengthQualifier of the length parameter in the map entry,
and here param="length{lengthQualifier=NONE}" should be used to change
the behaviour to having no qualifier at all in the DDL.

This is relevant for e.g. enumerations. Otherwise, each enumeration would need a separate map entry.