Remove xalan dependency, fix result error determination in unit tests
The xalan dependency was reported to have a security issue with critical
severity. Xalan-J itself does not seem to be well maintained anymore. In
fact, emails exchanged between xalan users/developers (on the xalan
mailinglist) suggest that the build process is not well understood and
has not been maintained, and that it would be a great effort to create a
new build and thus version of xalan (even with a security fix).
Therefore, we decided to remove the xalan dependency. As part of this
change, XML files are now written using a common utility method. The
encoding always is UTF-8.

Furthermore, errors and fatal errors from process log entries are now
taken into account when checking the result log for such errors when
determining the outcome of a unit test.

Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>