Documentation update - avoid footer with last updated info
The last updated date is misleading. With the build process of
ShapeChange, where the asciidoctor maven plugin is used to generate HTML
files from adoc, the last updated date simply represents the date that
the page was generated - not when a real content update occurred. By
removing the last updated date (the whole page footer, actually),
generated HTML pages can be compared more easily (actual content updates
can be spotted when a diff is performed).

Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>