fix creation of additional ontology models for default encoding
The default encoding of application schema packages (and the packages
contained therein) is to encode each package individually. That behavior
can be changed using `rule-owl-pkg-singleOntologyPerSchema`. Additional
ontology models need to be created in all cases - also in case that the
aforementioned conversion rule is NOT in effect AND for the case of the
application schema package itself. The latter was not the case, i.e.,
before this fix, additional ontology models were not created for the
application schema package; they were only created for the sub-packages.
The fix ensures that additional ontology models are created for all
cases.

fixes #431

Signed-off-by: Johannes Echterhoff <echterhoff@interactive-instruments.de>